+++ championships_handler.go
@@ -161,7 +161,7 @@ func (ch *ChampionshipsHandler) export(w http.ResponseWriter, r *http.Request) {
 		championship.OverridePassword = false
 	}
 
-	w.Header().Add("Content-Disposition", fmt.Sprintf("attachment; filename=%s.json", championship.Name))
+	w.Header().Add("Content-Disposition", fmt.Sprintf(`attachment; filename="%s.json"`, championship.Name))
 
 	enc := json.NewEncoder(w)
 	enc.SetIndent("", "  ")